pyproject.toml
[project]
name = "mlir-test"
version = "0.1.0"
description = "MLIR Blueprint Editor - Visual node-based MLIR programming"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mlir-python-bindings==20251208+fc92e4dec",
    "mlir-wheel==20251208+fc92e4dec",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "hypothesis>=6.100.0",
]

[tool.uv]
find-links = ["https://llvm.github.io/eudsl"]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]
